WE HAVE BASIC FUNCTIONALITY! The program now runs the basic package.
I added the basic program package, for our basic scenario. Some other minor alterations.

There are some caveats that need to be fixed:
The dataset constructor loads all of the relevant information, so it must read the file everytime.
There is no clipping, so sometimes the precision of the variables is off. This results in bad values (>1 || <0) for BasicAgent.
Potentially others that I can't think of, but those are the two.